Updated readme, added wakeup, and added easy to adjust variables to top of script.
Pantheon environments that are not live go into sleep mode.

If you ran the script against an environment that had not recently been visited, you would get a connection timeout, as the database server would be asleep.

Terminus has a command to wake up an environment. I added that to be fired if the environment is not live, before attempting to connect. This eliminates the timeout issue.

Variables have been added to the top of the connect.sh file, for ease of setup.